base_components/html_analyze/新建文件夹/requirements.txt
beautifulsoup4
python-dotenv
pika>=1.3.0
bleach[css]
esprima
lxml
boto3
requests